Don't strip units on "zero percent" values.
While we can safely strip units for zero values in other conditions,
"zero percent" needs to be retained in order to support CSS animations.
Rather than trying to figure out when/where that is, lets just not drop
units on zero percents at all.

We _will_ condense them where possible (e.g. "0000%" => "0%"), but we
don't drop the units.

Thanks to Isaac Montoya for continuing to poke me on getting this.