Use the `perldocker/perl-tester:latest` image to do our initial build.
This image already has DZil installed in it, so it cuts down on the time
needed to install all of our dependencies.

When we go to do the actual tests for the generated tarball, though, run
those on plain/naked Perl images to be sure that we've listed all of our
dependencies properly.