Rebuild internals, to avoid memory allocations for content buffers.
As we go through the process of tokenizing the document and collapsing
nodes, rather than allocating memory for the content buffers in each
node, just use pointers to point back into the original string we were
given.

We're not actually manipulating the contents of that string, so this is
safe to do, and *MUCH* faster.